# feat: Improve performance of substr by 84%