--- a/testdata/builtinSubStr_length_larger.jsonnet
+++ b/testdata/builtinSubStr_length_larger.jsonnet
@@ -1,0 +1,1 @@
+std.substr("hello", 0 , 10)