--- a/testdata/builtinSubStr_second_parameter_not_integer.jsonnet
+++ b/testdata/builtinSubStr_second_parameter_not_integer.jsonnet
@@ -1,0 +1,1 @@
+std.substr("hello", 1.2, 5)