--- a/testdata/builtinSubStr_second_parameter_not_number.jsonnet
+++ b/testdata/builtinSubStr_second_parameter_not_number.jsonnet
@@ -1,0 +1,1 @@
+std.substr("hello", "foo", 5)