--- a/testdata/builtinSubStr_start_larger_then_size.jsonnet
+++ b/testdata/builtinSubStr_start_larger_then_size.jsonnet
@@ -1,0 +1,1 @@
+std.substr("hello", 10, 5)