--- a/testdata/builtinSubStr_third_parameter_less_then_zero.golden
+++ b/testdata/builtinSubStr_third_parameter_less_then_zero.golden
@@ -1,0 +1,10 @@
+RUNTIME ERROR: substr third parameter should be greater than zero, got -1
+-------------------------------------------------
+	testdata/builtinSubStr_third_parameter_less_then_zero:1:1-27	builtin function <substr>
+
+std.substr("hello", 0, -1)
+
+-------------------------------------------------
+	During evaluation	
+
+