--- a/testdata/builtinSubStr_third_parameter_less_then_zero.jsonnet
+++ b/testdata/builtinSubStr_third_parameter_less_then_zero.jsonnet
@@ -1,0 +1,1 @@
+std.substr("hello", 0, -1)