--- a/testdata/builtinSubStr_third_parameter_not_integer.golden
+++ b/testdata/builtinSubStr_third_parameter_not_integer.golden
@@ -1,0 +1,10 @@
+RUNTIME ERROR: substr third parameter should be an integer, got 1.200000
+-------------------------------------------------
+	testdata/builtinSubStr_third_parameter_not_integer:1:1-28	builtin function <substr>
+
+std.substr("hello", 0, 1.2)
+
+-------------------------------------------------
+	During evaluation	
+
+