--- a/testdata/builtinSubStr_third_parameter_not_integer.jsonnet
+++ b/testdata/builtinSubStr_third_parameter_not_integer.jsonnet
@@ -1,0 +1,1 @@
+std.substr("hello", 0, 1.2)