--- a/testdata/builtinSubStr_third_parameter_not_number.jsonnet
+++ b/testdata/builtinSubStr_third_parameter_not_number.jsonnet
@@ -1,0 +1,1 @@
+std.substr("hello", 0, "foo")