--- a/testdata/builtinSubstr.jsonnet
+++ b/testdata/builtinSubstr.jsonnet
@@ -1,0 +1,1 @@
+std.substr("hello", 0, 5)