feat: Improve performance of substr by ~5x

This pulls in the implementation of substr into native Go instead of interpretted Jsonnet.

```
➜ ./benchmark.sh Builtin
Running Before Test... (10s)
Running After Test... (10s)
benchmark                       old ns/op     new ns/op     delta
Benchmark_Builtin_substr-16     96131271      14533163      -84.88%
```

part of #111