@ghostsquad 2020-01-25
Thank you for the great feedback. I'll get these changes in.